chore: Split Plume Legacy config into own util
